Exception Controll

+++ src/main/java/com/opbaquero/conexionaback/controllers/ProductRestController.java
@@ -2,8 +2,11 @@ package com.opbaquero.conexionaback.controllers;
 
 import com.opbaquero.conexionaback.models.entity.Account;
 import com.opbaquero.conexionaback.models.entity.Products;
+import com.opbaquero.conexionaback.models.exceptions.NotAllowedException;
 import com.opbaquero.conexionaback.models.service.interfaces.IAccountService;
 import com.opbaquero.conexionaback.models.service.interfaces.IProductService;
+import com.opbaquero.conexionaback.security.service.UserService;
+import com.opbaquero.conexionaback.utils.AuthorizationUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
@@ -27,6 +30,9 @@ public class ProductRestController {
     @Autowired
     private IProductService productService;
 
+    @Autowired
+    private AuthorizationUtils authorizationUtils;
+
     @GetMapping("/all")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public List<Products> all(){
@@ -45,6 +51,7 @@ public class ProductRestController {
 
     @PostMapping("/add/{accountId}")
     @ResponseStatus(HttpStatus.CREATED)
+    @PreAuthorize("hasRole('ROLE_GESTOR')")
     public ResponseEntity<?> createProduct(@PathVariable (value = "accountId")UUID accountId, @RequestBody Products products){
         Account account = accountService.findOne(accountId);
         Map<String, Object> response = new HashMap<>();
@@ -60,20 +67,28 @@ public class ProductRestController {
     }
 
     @DeleteMapping("/add/{productId}")
+    @PreAuthorize("hasRole('ROLE_GESTOR')")
     public ResponseEntity<?> delete(@PathVariable(value = "productId") UUID productId){
         Map<String, Object> response = new HashMap<>();
         try{
-            productService.delete(productId);
+            if(this.authorizationUtils.checkAccountsProducts(productId)){
+                productService.delete(productId);
+            }
         }catch (DataAccessException e){
             response.put("message", "Error deleting product from database");
             response.put("error", e.getMessage().concat(":").concat(e.getMostSpecificCause().getMessage()));
             return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+        }catch (NotAllowedException e){
+            response.put("message", "Product not found in account");
+            response.put("error", e.getMessage().concat(":").concat(e.getMessage()));
+            return new ResponseEntity<Map<String, Object>>(response, HttpStatus.FORBIDDEN);
         }
         response.put("message", "Product deleted");
         return new ResponseEntity<Map<String, Object>>(response, HttpStatus.OK);
     }
 
     @PutMapping("/add/{productId}")
+    @PreAuthorize("hasRole('ROLE_GESTOR')")
     public ResponseEntity<?> updateProduct(@RequestBody Products products, @PathVariable (value = "productId") UUID productId){
         Products currentProduct = productService.findOne(productId);
         Products updateProduct = null;
